refactor(tui): split interaction history tests + layout helpers under 400-line cap
CI File Size Lint failed: history.rs (507) and mod.rs (414) exceeded the
400-line limit after the #987 card work.

- Extract history.rs inline test module to history_tests.rs (#[path]).
- Extract layout/scroll helpers (HEADER_HEIGHT, INPUT_HEIGHT, inset_x,
  effective_offset) from mod.rs to a new layout.rs.

No behavior change. history.rs 215, mod.rs 381, layout.rs 40.

Refs #987
