+++ src/tui/screens/interaction/history.rs
@@ -211,297 +211,5 @@ fn starter_hint<'a>(theme: &Theme, issue_number: u64, issue_title: &'a str) -> P
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::tui::theme::Theme;
-    use chrono::{TimeZone, Utc};
-
-    fn turn(role: TurnRole, content: &str, streaming: bool) -> TurnRecord {
-        let at = Utc.with_ymd_and_hms(2026, 1, 1, 9, 0, 0).unwrap();
-        TurnRecord {
-            role,
-            content: content.to_string(),
-            started_at: at,
-            finished_at: if streaming { None } else { Some(at) },
-        }
-    }
-
-    /// Collect the plain Unicode text of a line, stripping all style.
-    fn line_text(line: &Line<'_>) -> String {
-        line.spans.iter().map(|s| s.content.as_ref()).collect()
-    }
-
-    fn count_prefix(lines: &[Line<'_>], prefix: &str) -> usize {
-        lines
-            .iter()
-            .filter(|l| line_text(l).starts_with(prefix))
-            .count()
-    }
-
-    // -----------------------------------------------------------------------
-    // role_color mapping (unchanged behavior)
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn role_color_three_roles_produce_three_distinct_colors() {
-        let theme = Theme::dark();
-        let user = role_color(TurnRole::User, &theme);
-        let agent = role_color(TurnRole::Agent, &theme);
-        let system = role_color(TurnRole::System, &theme);
-        assert_ne!(user, agent, "User and Agent must differ");
-        assert_ne!(user, system, "User and System must differ");
-        assert_ne!(agent, system, "Agent and System must differ");
-    }
-
-    #[test]
-    fn role_color_user_maps_to_accent_info() {
-        let theme = Theme::dark();
-        assert_eq!(role_color(TurnRole::User, &theme), theme.accent_info);
-    }
-
-    #[test]
-    fn role_color_agent_maps_to_accent_success() {
-        let theme = Theme::dark();
-        assert_eq!(role_color(TurnRole::Agent, &theme), theme.accent_success);
-    }
-
-    #[test]
-    fn role_color_system_maps_to_text_secondary() {
-        let theme = Theme::dark();
-        assert_eq!(role_color(TurnRole::System, &theme), theme.text_secondary);
-    }
-
-    // -----------------------------------------------------------------------
-    // Card structure
-    // -----------------------------------------------------------------------
-
-    #[test]
-    fn build_lines_settled_card_has_header_body_footer_and_blank() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::User, "hi", false)];
-        let lines = build_lines(&history, &theme, 80);
-        assert_eq!(lines.len(), 4, "header + body + footer + blank");
-        assert!(line_text(&lines[0]).starts_with("╭─ you · 09:00"));
-        assert!(line_text(&lines[0]).ends_with('╮'));
-        assert!(line_text(&lines[1]).starts_with("│ "));
-        assert!(line_text(&lines[1]).ends_with(" │"));
-        assert!(line_text(&lines[2]).starts_with('╰'));
-        assert!(line_text(&lines[2]).ends_with('╯'));
-        assert_eq!(line_text(&lines[3]).trim(), "");
-    }
-
-    #[test]
-    fn build_lines_streaming_card_has_no_footer() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::Agent, "Working on it", true)];
-        let lines = build_lines(&history, &theme, 80);
-        assert_eq!(
-            count_prefix(&lines, "╰"),
-            0,
-            "streaming turn must omit footer"
-        );
-        assert!(
-            line_text(&lines[0]).contains('…'),
-            "streaming header carries …"
-        );
-    }
-
-    #[test]
-    fn build_lines_header_contains_role_and_time() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::User, "hello", false)];
-        let lines = build_lines(&history, &theme, 80);
-        let header = line_text(&lines[0]);
-        assert!(header.contains("you · 09:00"), "header: {header}");
-        assert!(header.starts_with("╭─"));
-        assert!(header.ends_with('╮'));
-    }
-
-    #[test]
-    fn build_lines_header_role_words() {
-        let theme = Theme::dark();
-        for (role, word) in [
-            (TurnRole::User, "you"),
-            (TurnRole::Agent, "agent"),
-            (TurnRole::System, "sys"),
-        ] {
-            let history = vec![turn(role, "x", false)];
-            let lines = build_lines(&history, &theme, 80);
-            let header = line_text(&lines[0]);
-            assert!(
-                header.contains(&format!("{word} · 09:00")),
-                "{role:?} header: {header}"
-            );
-            assert!(!header.contains('▸'), "no legacy arrow glyph");
-        }
-    }
-
-    #[test]
-    fn build_lines_header_span_uses_role_color() {
-        let theme = Theme::dark();
-        for (role, color) in [
-            (TurnRole::User, theme.accent_info),
-            (TurnRole::Agent, theme.accent_success),
-            (TurnRole::System, theme.text_secondary),
-        ] {
-            let history = vec![turn(role, "x", false)];
-            let lines = build_lines(&history, &theme, 80);
-            assert!(
-                lines[0].spans.iter().any(|s| s.style.fg == Some(color)),
-                "{role:?} header must carry its border color"
-            );
-        }
-    }
-
-    #[test]
-    fn build_lines_body_lines_wrapped_in_border_pipes() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::Agent, "- alpha\n- beta", false)];
-        let lines = build_lines(&history, &theme, 80);
-        let body: Vec<_> = lines
-            .iter()
-            .filter(|l| line_text(l).starts_with("│ "))
-            .collect();
-        assert!(body.len() >= 2, "multiline markdown body spans >= 2 rows");
-        for line in body {
-            let t = line_text(line);
-            assert!(t.starts_with("│ "), "body row left gutter: {t}");
-            assert!(t.ends_with(" │"), "body row right gutter: {t}");
-        }
-    }
-
-    #[test]
-    fn build_lines_body_preserves_markdown_styling() {
-        let theme = Theme::dark();
-        // Bold markdown must survive into a boxed body span.
-        let history = vec![turn(TurnRole::Agent, "**bold**", false)];
-        let lines = build_lines(&history, &theme, 80);
-        let has_bold = lines.iter().flat_map(|l| l.spans.iter()).any(|s| {
-            s.style
-                .add_modifier
-                .contains(ratatui::style::Modifier::BOLD)
-        });
-        assert!(has_bold, "render_markdown bold styling must be preserved");
-    }
-
-    #[test]
-    fn build_lines_empty_content_body_has_one_padded_line() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::User, "", false)];
-        let lines = build_lines(&history, &theme, 80);
-        assert_eq!(lines.len(), 4, "header + 1 empty body + footer + blank");
-        assert!(line_text(&lines[1]).starts_with('│'));
-        assert!(line_text(&lines[1]).ends_with('│'));
-    }
-
-    #[test]
-    fn build_lines_content_truncated_to_inner_width() {
-        let theme = Theme::dark();
-        let width: u16 = 40;
-        let history = vec![turn(TurnRole::User, &"x".repeat(120), false)];
-        let lines = build_lines(&history, &theme, width);
-        for line in &lines {
-            assert!(
-                line.width() <= width as usize,
-                "no line may exceed card width {width}: got {} ({:?})",
-                line.width(),
-                line_text(line)
-            );
-        }
-    }
-
-    #[test]
-    fn build_lines_narrow_width_does_not_panic() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::User, "hello world", false)];
-        let lines = build_lines(&history, &theme, 40);
-        assert!(line_text(&lines[0]).starts_with("╭─"));
-        assert!(line_text(&lines[0]).ends_with('╮'));
-    }
-
-    #[test]
-    fn build_lines_blank_separator_between_two_cards() {
-        let theme = Theme::dark();
-        let history = vec![
-            turn(TurnRole::User, "hi", false),
-            turn(TurnRole::Agent, "ok", false),
-        ];
-        let lines = build_lines(&history, &theme, 80);
-        assert_eq!(count_prefix(&lines, "╭─ you"), 1);
-        assert_eq!(count_prefix(&lines, "╭─ agent"), 1);
-        assert_eq!(count_prefix(&lines, "╰"), 2, "two footers");
-        // A blank line sits before the second card's header.
-        let agent_idx = lines
-            .iter()
-            .position(|l| line_text(l).starts_with("╭─ agent"))
-            .expect("agent header present");
-        assert_eq!(
-            line_text(&lines[agent_idx - 1]).trim(),
-            "",
-            "blank separator precedes the next card"
-        );
-    }
-
-    #[test]
-    fn build_lines_streaming_ellipsis_in_header_not_body() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::Agent, "content here", true)];
-        let lines = build_lines(&history, &theme, 80);
-        assert!(line_text(&lines[0]).contains('…'), "header carries …");
-        let body = line_text(&lines[1]);
-        assert!(!body.trim_end().ends_with('…'), "body must not end with …");
-    }
-
-    #[test]
-    fn build_lines_two_single_line_turns_emit_two_cards() {
-        let theme = Theme::dark();
-        let history = vec![
-            turn(TurnRole::User, "hi", false),
-            turn(TurnRole::Agent, "ok", false),
-        ];
-        let lines = build_lines(&history, &theme, 80);
-        // Each settled single-line turn => header + body + footer + blank.
-        assert_eq!(lines.len(), 8);
-        assert!(line_text(&lines[0]).starts_with("╭─ you"));
-        assert!(line_text(&lines[4]).starts_with("╭─ agent"));
-    }
-
-    #[test]
-    fn build_lines_multiline_content_emits_multiple_body_rows() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::Agent, "- a\n- b\n- c", false)];
-        let lines = build_lines(&history, &theme, 80);
-        let body_rows = count_prefix(&lines, "│ ");
-        assert!(body_rows >= 3, "three list items => >= 3 body rows");
-        assert_eq!(count_prefix(&lines, "╭─"), 1, "single card header");
-        assert_eq!(count_prefix(&lines, "╰"), 1, "single card footer");
-    }
-
-    #[test]
-    fn visual_total_counts_card_rows() {
-        let theme = Theme::dark();
-        let history = vec![turn(TurnRole::User, "hi", false)];
-        let total = visual_total(&history, &theme, 80);
-        assert!(total >= 4, "card occupies at least 4 rows: {total}");
-    }
-
-    #[test]
-    fn visual_total_equals_built_line_count() {
-        // Cards are pre-boxed to exactly `width`, so no line wraps: the visual
-        // row count is just the number of built lines. This keeps scroll math
-        // a straight 1:1 with the flat line vector.
-        let theme = Theme::dark();
-        let history = vec![
-            turn(TurnRole::Agent, &"x".repeat(100), false),
-            turn(TurnRole::User, "short", false),
-        ];
-        for width in [40u16, 80, 200] {
-            let total = visual_total(&history, &theme, width);
-            let built = build_lines(&history, &theme, width).len();
-            assert_eq!(
-                total, built,
-                "visual_total must match line count at {width}"
-            );
-        }
-    }
-}
+#[path = "history_tests.rs"]
+mod tests;
